# openapi.yaml
openapi: 3.0.3
info:
  title: Portfolio Desk API
  version: 1.0.0
  description: |
    Multi-tenant office management platform API. Handles offices, leases, maintenance tickets,
    vendors, HVAC contracts, billing, and analytics.

    ## Authentication
    All protected endpoints require a JWT bearer token obtained from `/api/v1/auth/login`.
    API keys (prefixed `om_`) may be used as an alternative bearer credential.
servers:
  - url: /api/v1
    description: Primary API
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/login or an API key prefixed with `om_`

  # -------------------------------------------------------------------------
  # Reusable schemas
  # -------------------------------------------------------------------------
  schemas:
...
    CheckoutRequest:
      type: object
      required: [plan]
      properties:
        plan:
          type: string
          enum: [pro, enterprise]
    CheckoutResponse:
      type: object
      properties:
        checkout_url: { type: string, format: uri }
    PortalResponse:
      type: object
      properties:
        portal_url: { type: string, format: uri }

    # --- Dashboard ----------------------------------------------------------
...
security:
  - BearerAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /billing/checkout:
    post:
      tags: [Billing]
      summary: Create Stripe Checkout session (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CheckoutRequest' }
      responses:
        '200':
          description: Checkout URL
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CheckoutResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /billing/portal:
    post:
      tags: [Billing]
      summary: Create Stripe Customer Portal session (admin only)
      responses:
        '200':
          description: Portal URL
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortalResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /billing/webhooks:
    post:
      tags: [Billing]
      summary: Stripe webhook receiver (HMAC verified, public)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: Received
          content:
            application/json:
              schema:
                type: object
                properties:
                  received: { type: boolean }

  # ==========================================================================
  # API KEYS
  # ==========================================================================
...